rest-backend/requirements.txt
Django
djangorestframework
markdown
django-filter
coreapi
django-cors-headers
psycopg2-binary
paho-mqtt